+++ src/components/common/CorpInfoCarousel.vue
